+++ app/src/components/Toolbar/fragments/ToolbarButton.tsx
@@ -2,11 +2,10 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 import { styled, Tooltip, ThemeProvider } from '@mui/material';
-import { ArrowDownIcon } from '@opentalk/common';
+import { ArrowDownIcon, IconButton } from '@opentalk/common';
 import { MouseEvent, ReactNode } from 'react';
 
 import { createOpenTalkTheme } from '../../../assets/themes/opentalk';
-import IconButton from '../../../commonComponents/IconButton';
 
 interface IButtonProps {
   isActive?: boolean;